chore: Rename v2 table files
