+++ packages/table/src/tableUtils.ts
